Factor out submodules from Cluster module
