+++ lib/local-cluster/lib/Cardano/Wallet/Launch/Cluster/PoolMetadataServer.hs
@@ -1,0 +1,66 @@
+module Cardano.Wallet.Launch.Cluster.PoolMetadataServer
+    ( PoolMetadataServer (..)
+    , withPoolMetadataServer
+    )
+where
+
+import Prelude
+
+import Cardano.BM.Tracing
+    ( Tracer
+    , traceWith
+    )
+import Cardano.Wallet.Launch.Cluster.Logging
+    ( ClusterLog (MsgRegisteringPoolMetadata)
+    )
+import Crypto.Hash.Extra
+    ( blake2b256
+    )
+import Data.ByteArray.Encoding
+    ( Base (Base16)
+    , convertToBase
+    )
+import System.Directory
+    ( createDirectoryIfMissing
+    )
+import System.FilePath
+    ( (</>)
+    )
+import Test.Utils.StaticServer
+    ( withStaticServer
+    )
+
+import qualified Data.Aeson as Aeson
+import qualified Data.ByteString.Char8 as B8
+import qualified Data.ByteString.Lazy as BL
+import qualified Data.ByteString.Lazy.Char8 as BL8
+
+data PoolMetadataServer = PoolMetadataServer
+    { registerMetadataForPoolIndex :: Int -> Aeson.Value -> IO ()
+    , urlFromPoolIndex :: Int -> String
+    }
+
+withPoolMetadataServer
+    :: Tracer IO ClusterLog
+    -> FilePath
+    -> (PoolMetadataServer -> IO a)
+    -> IO a
+withPoolMetadataServer tr dir action = do
+    let metadir = dir </> "pool-metadata"
+    createDirectoryIfMissing False metadir
+    withStaticServer metadir $ \baseURL -> do
+        let _urlFromPoolIndex i = baseURL </> metadataFileName i
+        action PoolMetadataServer
+            { registerMetadataForPoolIndex = \i metadata -> do
+                let metadataBytes = Aeson.encode metadata
+                BL8.writeFile (metadir </> (metadataFileName i)) metadataBytes
+                let hash = blake2b256 (BL.toStrict metadataBytes)
+                traceWith tr
+                    $ MsgRegisteringPoolMetadata
+                        (_urlFromPoolIndex i)
+                        (B8.unpack $ convertToBase Base16 hash)
+            , urlFromPoolIndex = _urlFromPoolIndex
+            }
+  where
+    metadataFileName :: Int -> FilePath
+    metadataFileName i = show i <> ".json"
